requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0
pillow>=9.0.0
opencv-python>=4.7.0

# Machine Learning
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
tqdm>=4.64.0

# Experiment Tracking (optional)
wandb>=0.13.0

# Jupyter/Colab
ipykernel>=6.0.0
ipywidgets>=8.0.0
